scripts/requirements.txt
# Dependencias para scripts de facturación electrónica DIANA
requests==2.31.0
cryptography==41.0.7
lxml==4.9.3
zeep==4.2.1
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
click==8.1.7
rich==13.7.0
xmltodict==0.13.0
pycryptodome==3.19.0
certifi==2023.11.17
urllib3==2.1.0
